Add delete_role functionality.

+++ src/discord/commands/leaderboard/lb.remove.ts
@@ -1,4 +1,4 @@
-import { CommandInteraction } from "discord.js";
+import { CommandInteraction, User } from "discord.js";
 import { DB } from "../../../db";
 import { Leaderboard, TrackedLeaderboard } from "../../../db/models";
 import UserError from "../../UserError";
@@ -10,6 +10,7 @@ export async function remove(interaction: CommandInteraction) {
 	const lbRepo = DB.getRepository(Leaderboard);
 	
 	const lb_id = interaction.options.getInteger('leaderboard')!;
+	const delete_role = interaction.options.getBoolean('delete_role');
 
 	const lbCrit = { where: { lb_id }, relations: { trackedLeaderboards: true } };
 	const lb = await lbRepo.findOne(lbCrit);
@@ -18,6 +19,12 @@ export async function remove(interaction: CommandInteraction) {
 	const tlbCriteria = { where: { guild_id: interaction.guildId!, lb_id } }
 	const tlb = await tlbRepo.findOne(tlbCriteria);
 	if(!tlb) throw new UserError(`This guild does not track the given leaderboard.`);
+
+	if(delete_role) {
+		const boardsWithRole = await tlbRepo.find({ where: { role_id: tlb.role_id } });
+		if(boardsWithRole.length > 1) throw new UserError(`Error: More than one leaderboard use the role for this board.`);
+		await interaction.guild!.roles.delete(tlb.role_id);
+	}
 
 	// delete the tracking
 	await tlbRepo.delete(tlbCriteria.where);
